Updated lib and code of changed methods
changed return value of long to int in some methods

+++ src/org/sosy_lab/java_smt/solvers/boolector/BoolectorAbstractProver.java
@@ -83,7 +83,7 @@ abstract class BoolectorAbstractProver<T> extends AbstractProverWithAllSat<T> {
   public boolean isUnsat() throws SolverException {
     Preconditions.checkState(!closed);
     wasLastSatCheckSat = false;
-    final int result = (int) BtorJNI.boolector_sat(btor);
+    final int result = BtorJNI.boolector_sat(btor);
     if (BtorSolverResult.swigToEnum(result) == BtorSolverResult.BTOR_RESULT_SAT) {
       wasLastSatCheckSat = true;
       return false;
